Move crate::network::peer_manager to crate::peer
This requires that the SendRequest and NetworkSender
are updated to pub crate. these structs will be moved to
crate::peer.
