+++ libsplinter/src/peer/error.rs
